Add files via upload
Includes previous 1-4 steps plus includes actually getting the waves - DC, Sine, sqaure, triagular on the analyser.
For this, a pre-fed 1V, 1khz wave is stored in an array and then, according to the required frequency, the delphi variable - representing the index of the pre-fed array is increased in terms of the frequency. 

Also, DAC gives 0-2.5V output, hence, using op-amps, it is mapped from -5 to +5. Hence to get the desired voltage (from -5 to 5) the value fed into the DAV is calculated. E.G. for DAC input 0, the output of opamp is -5V. Also, some calibration is needed as the voltage won't be exactly -5V for 0 DAC value.