Add files via upload
Step 9,10 includes increasing frequency of a wave from f1 to f2 frequency in steps. This is done with the help of inbuilt ADC and sweep function. The sweep function starts with f1 frequency. If the frequency is less than 1kHz, it starts with a slow ADC. It keeps on increasing the frequency in the power of 2 i.e freq+2^i, where i =0,1,..N and N depends on f2 frequency. Once the frequency is past 1Khz, the ADC starts fast ADC. Hence, the ouput wave will sweep from frequency f1 to f2.
Slow ADC - voltage is converted every 3 sec Fast ADC - voltage is converted every 0.5 sec.